Add base language model based on feed forward network only
